Only drop None's on the way to Dynamo
The last two changes meant we dropped None everywhere, but in the case
of Schematics it has a ``serialize_when_none`` parameter that we should
be honoring in the schema design.

None's are now dropped in the table only on their way to Dynamo.

The hardest part of this was figuring out how to mock it out to test
it...